contact do not longer need to active body when destroyed
Starting with Newton 3.12 and 3.13 contact destruction is no longer
controlled by the broadphase, instead they garbage collected. when they
are no longer active because their aabb do not overlap.

This offer several advantages for using contact joint for triggers, and
also activation deactivation do not have to wake up entire island just
because they were close to it but not touching the island