think I fix the bug with the  brad phase optimization
added the tow bradphases.
the default with world well for all scene
the persistent which optimized for scenes where the number of static
bodies is much larger that dynamics objects , the brad phase is  little
slower for mix scene but some new game are suing 10's of tousands  of
static bodies and only few hundred dymancis, and this broad phase do no
spent no time of static objects so the time gain there far compensated.

also remove the profiler stiff form the engine, I will try to add
visual studio concurrency markes with seems very nice and it is free.