added Broad Phase aggregate
Aggregate are similar to and scene compound collision, but the group
full set of dynamics bodies with mass.
aggregates are  special broad phase node that can group a number of
rigid bodies the application know will always be close together.

A typical case with ne a Vehicle when the app know the all the body part
are no self collidable this can save lot of calculations
another case would be a rag ball that connect many body part who are
collidable but the are so close that the bound of one bound is almost
identical to the bound of the whole, this saves lots of broad phase
iterations

another example is a City of a building with many room and populated
with many dynamics bodies, since most bodies inside a room will be a
rest (sleeping)  the application can encapsulate these bodies in a
aggregate and save thousands of calculation, until some other object
intact with the room

Aggregate provide an interface to the application to organize the scene
by grouping bodies the way if fix better.