more work on vehicle,
vehicle still have a clear tendency to pull to the left, this make not
sense since everything has symmetry, and the veering is too large to  be
accounted by numerical accuracy.
I need to see a test and plot all the solver values to make sure it does
not comes from there.