ok the QR factorization was a waste of time
It is true QR is more efficient that, that a Jacobi Givens rotation ,
who ever for a 3 x 3 moment matrix, when using first guess Jacobi
converge in one of tow iteration while QR still takes more than three.
There are simple not enough off diagonal zeros in a 3 x 3 matrix for the
overhead of household to be effective.  I state with Jacobi for now.