experimenting with a Lagrangian Dynamics solver for the vehicle.
Ok I accept that the velocities range and the masses ratios involved of
a realistic vehicle are too high for a iterative solver, I am try to
solve the problem using Lagragian dynamics solver in a generalize space,

so that I can calculate the vehicle forces and apply the to the vehicle
body as external forces.
A generalized space solution is no easy to tune by at least solve the
numerical error problems, which for what I can see has no way to solve.