improvement on real differential joint
It turned out that the mathematic of a differential joint is in fact and
over determined system,
This is why I have so much problem try to make one for a high
performance car.
It is possible to make a half differential by using only one satellite
gear and the avoid the singularities, however is has the side effect
that that the centripetal force generate by the satellite gear does not
cancel.
what this mean, is that we much chose a very small mass object for the
differential gear, or eh gear can only rotate a relatively low speed.
Real differential do not have this problem because the first the
satellite gears are very small mass and they are used in symmetric
pairs.

for a tank this implementation is OK and is actually quiet nice.