added meg flop per secunds counter
I am getting an anemic 4 giga flop per secund, thsi is very disappointing is seem tha usin simple scaler operation is actually better than using SSE or AVX.

It also appears that fmul add is actually slower that using mul and add separate.
Maybe this is because the loop is dominated by memory banwidth, but as it stand now there is not benefit from using AVX2, if anything seems slower.