start removing data struture of array form paralle solver (WIP)
having the paralle solve use stucture of array add an extra opver head to the solve and more imprtant it prevent the joint to be part if teh solver, so I will remove it.

also when moving to GPY ther is no nee for traposing teh data, since GPU have that dat gathreing and scatering intruction.