Removed customized code for install directories
Rather than using the homebrewed directories for file install locations,
it now uses the default bin/lib/include dirs and the typical
CMAKE_INSTALL_PREFIX to determine install locations. To emulate the older
behaviour, the CMAKE_INSTALL_PREFIX on Windows defaults to the 'build'
dir.

This ultimately allows for proper customization of the install location
on Windows through the CMAKE_INSTALL_PREFIX variable.