Greatly reduced dependency graph of libraries
This reduces the inter-library depndencies to the actual requirements,
also reducing extraneous linking commands. The same also applies for
many of the include directories, that are not attatched to the targets
themselves, and thus their properties will be carried with them to other
targets that use them.

As well, the core 'newton' library can be generated on its own if
requested via the NEWTON_BUILD_CORE_ONLY CMake variable.