re added implicit surface concat point normal,
for tire the generic collision is not precise enough. tiny penetration errors cause the surface velocity to change enought that the slip value keeps changing by small amount. Since the system is stiff this small change translate to significan tire forces.

The implict contact project the concats to the tire surface and this guarantee uniform contact point on the surface.