start working on setting up a very simpler self balacing dynamics biped
previus demos where too complex and lost sight of the objective.
I am now trying very simple set up hoping people have less difficulties following the developent.
the model is try to make a self balacing biped with a simple hip, torso for center of mass counte weight, and two legs. teh model in total will have 18 degree of freedoms. 3 for each legs, 1 for each shim, and 2 for each foot = 12 dpf for the limbs plus 6 for the hip. making an 16 dof model.
my stimation is that this is the simplest full articulated 3d biped model that can be made.