more work on setting bipe model.
I now solved the bug with the twist angle and the bipe seem stable, is was just a bug on the joint pin.
the system seem quite stable, so I can more on to work on teh sefl balacing controller.
whoever I knwo that once the leg start to move the conatc will start moving around and I at that ppoint I have to impemnet the soft contact.
I call a soft contact a contact that absorve kenetic energy by letting the shape penetration a litle simulation a plastic body. penetartion is resolve with a small fix velocity.
by having the shape be in penetration, teh constact will be more stable.
at least that's my hypothesis.