play a litel with stability and behavior. (WIP)
limb forces do not really need to be so strong, in fact by experimentaion,
it seems that a factor if the body weight is quite sufficient.
this make sense, because a human of any mamal can no really lift its own body weight many time.
ok move one with more tests.