made lock mutex free the one an only thread pool method
using thread wait seems fart more expensive than just using thread yield during the physics update and the semaphore at the beginning and end of the physics update.
still the very first signaling of the thread pool seems to cost 0.3 ms, and I can't explain why this cost is so high.  but it seems to be the fix cost of having a thread pool of the windows thread scheduler.