working on debri connectivity, using geometric approach
calculation the edge connectivity can be done analytical from the voronoi diagram,
but is has the problem that it will reports all the adjacent convex that share a feature.

stuff like convex sharing edge-point, edge-edge, point-point, face-edge and so on.
we only need face-face features pair and filtering the extra feature is as complex as using a geometrical approach.