make defualt solve not use large local variable array.
need to this because the default solve is teh model for Opencl.
since the solve use an array of local variable that is larger that is about 3.5 k, thsi si too large for solve a workgrop of 64, 128, or 256 element sine teh size will be way large that the teh share memory in a compute unit.