making opencl test
body integration only,
opencl shows that there is lots of optimization work to do even for the cpu solvers.
sorting is the major factor.  that nee to be replace with radix.
also need flag for objects that do no need call back, ex for most bodies gravity is constant, the eliminate few passes.