made must containers sub class of dClassAlloc
this is necessary because because garbage collected languages  and tools like swig
all object are allocated,  but the standard new and delete operators do not guarantee that memory will be 32 byte aligned.
an example is the python wrapper,  make a dVector crash because the vector is allocated using new and was no 16 byte aligned.
adding a global new fixe the problem, but the real solution was to make dVector derive from dClassAlloc. same goes for all other containers.