using atomic for cpu but not for GPU
from here the GPU and CPU solver start to diverge.
the algorithm for making the solver that use hundred of thousand of thread use too much memory of it was gong to be anomic free.  but on cpu atomics  are fast if they are no on the same cache line.

GPU solver will using a atomic lock free method and if it shows to be as effecient as the defaul solve on CPU, them we ported to teh CPU, but for now the solve will diverge.