stress testing lock free solver
we have a much better yield than in multithreaded than when using atomic writes.
now avx2 is literately 5 time faster on a 6 core using 6 thread than is is in single thread. This mean the solver is much faster since it is the only part that was optimized,  other part still has the same cost.