removed early out from opencl solver
early out make the opencl solver twice slower. but that fine since we caon on GPU higher floatps/ secund
but In will be interesting to check if it does eh same to the avx  and see
I suspect that it does something to the sse, but no effect on avx since is hard to have 8 joints converge to zero in on call. but I need to test it