using native float8 instead of float4,
some GPU claim to have memory coalesced read and writes, but this is only for local memory, I am using explicit 256 read and writes.
my  hope is that by using explicitly  256 read and writers instead of realizing on the compiler to optimize those memory transactions, that we better optimized the full  256 bus.