start replacing array of pointer with arrays of indices in the solver side
it is too difficult and error prompt to write a opencl solve with array of pointer and change them at the same time.
my first try I made few mistakes, so it is better to do in two steps.
first replace pointer with indices and test that the solver works the same way.
then start converts cpu kernels to GPU kernels.