it seems that most preconditioners for  conjugate gradient descend are not much better that plain Jacibi
Incomplete Cholesky seem to be of academic value, since is very unlike to have a LLt factorization with many near zero elements.

this leave me with up and down Gauss sidle but I try that in 1.xx and I remember the cost is at a minimum two extra matrix multiply so in end up with three matrix multiply in the inner loop for a mere one of two few passes, and in some case the converge rate was no great.
so It seems I will state with Jacobi Preconditioner.