I now have the new parallel island working
the new algorism generation the same sequence of bodies and Joints as the sequential one,  but does not generate individual islands.
The new method can not put a entire island to sleep unless all bodies goes to rest,
so now I nee to implement the progressive method, the weight the bodies acceleration threshold based of there location in the island.

but all in all this algorithm is much better faster when using multicore