ok I believe I found the bug that was cause all these problems
it turned out I did have a race condition in InitWeight
I was writing values that are supposed to be monotonically increasing to the joint weight, because the joints are sort by counts, however when running muticore two thread can write out of order. and the last write could be smaller.  and that was the huge bug that hunted me for month.

on teh bright side, I did find many fixes that do make the engine more streamlined.