enabling avx2 solver.
happy to say that with teh lock free algorithm for teh firs time I see avx2
beating by a substantial margin the sse solver
for comparison a 30 x 30 pyramid running on a icore7 7700  four threads
the default solver takes 11 to 12 ms per frame to bring to rest.
sse solver takes 7 to 8 ms per frame to bring to rest.
avx2  solver takes 4 to 5 ms per frame to bring to rest.

avx2 is almost twice as fast, so intel was right. avx2 really yield double or more floats but it is really hard to get that kind of performance.
you have to have algorithms that are almost no memory bound, very few branches and to atomics operations.