writing fluid solve for CPU,
need to separate the CPU and GPU code.
in CPU a 64 x 64 x 64 block of particle is equivalent to 256,000 particles.

the intermediate buffer is 8 entries per particles so that's 2,100,000 hash
of 64 bit each which is a 12 meg buffer.
Radix sort need two of those buffers, so that 32 meg of byte to be move randomly using counting sort. is simply blows out the cache.
so no even a modern cpu will be able to copy with that.

for that reason the cpu solver will use a 2d beam sorting so that the memory will be cut by more than half. so far seem very good, but I have no written the pair generator yet. to compare

I have not idea why GPU has such high bandwidth, and why we can no have that with CPUs.
maybe AMD idea of eth 3D cache is what we wait for.