using nd name space to v-hacd
the modified version of v-hacd, does not use opencel, openmp or simd.
is in fact faster and generated better result that the current github version.

since the github is not accepting my contribution, I have to add a unique name space to the library so that is does no create linking conflict with othe engine using the same library.

`