renamed ndJointHingeActuator to  ndJointHingePd
proportional derivative motors are  are much better that actuator because they can be pressure sensitive.
this mean that for example we can make a gripper that can get objects of different weights by setting the penetration error, while a actuator can onle apply the max friction force.

the only problem is that PD rows are harder to solve because they can have more than 6 dof, so the cost more in solve time when the hit a limit.