more work in double hinge joint
this is one of the most difficult joint to archive, this is because is very hard to calculate the counter toque gyro, specifically when the angle velocity is variable like in teh case of the oscillator.

in teh pass I always use the trick of make the inerta spherical, but that bring the problem of slow or extreme unstable reaction sine the inertia of teh smaller axis is teh one that dominate the motion. and the spin never align to that axis as it should.
anyway so far so good.

still not done yet, the spherical joint with three motors is even harder