more impulse respose testing of the diffrent chains.
there seems to be a phase error larger that I expected.
this is normal for proportional derivative systems,  but I still do not see why there is a diffrent on the y axis when only the x axis is changing.
I think is the non uniform inertia, so I will check that later.
also seem to be inverted for front and rear legs.