testing swivel angle control on the default plane.
joint seem to fallow almost perfect.
ther is still a small error on that make teh joint veer of teh plane, but that's numerical integration and high gyro inretia. the onel way to reduce that error in by taking smaller time step.

so far the dominant error is the phase angle of the integral derivative, and that is expected.

now need to test animating teh swivel angle. and also the selection of the prefer solution with reaching a point with multiples legal IK solutions.