using a virtual feet limb
controlling a 4 dof chain with a 6 dof effector makes an non holonomic over determined system.
This required to make kinematic relation on the effector so some two angles are function of the relative position cancelling two rows.
this will cause one of the row to be iterated and limited by the solver, but the calculation of the requires some extra angular velocities generated by the rotation swivel axis which the joint does not handle and the motion follows a different path that expected.

a solution is to make the IK chain a full 6 dof by adding a virtual feet body with teh tow extra rotation degree of freedom,
thsi way teh effector can solve all dof as a long and teh effector position and rotation is such that not dof is lost.
It is teh job of teh effector controller to always place the effector in valid workspace positions. (a task for AI, or state machine)