start process of eliminating island
island construction and testing is a n log (n) that is very hard to parallelize.
it is better to have  one or more extra pass whi is linear, so that residual accelration is reduce more and eliminate the island loop.
this is specially important for GPU and very large scenes.
It is still a investigation, it may be that teh PC could benefit while GPU do not not need it.