ok it seems there is a bug in the ndScene copy constructor
so chnge solver mode will crash until I find the bug.
The nd Scene is a complex graph that can no be eassy recreated, it has to be reconstructed correctly form teh current and it seems I am deleting or not rebasing some buffers correctly.
also faster sleep check by the connectivity, not need to count joint and contacts.