debugging crash bug when chnage scene.
It seem that there an object that is not releasing it reference to another object when deleting and some how use it form teh free list.
this is really bad, and I nee to add some debug code to stress test the free list.
It does no happen if the scene does no changes, but it think the bug is still there just take longer to happens.