sunstancial improvement in sph particles.
reduce memory band with by doing two passes, on et count  grid, and one to generate them in place.
before it was making a large array with duplicates in a temp set, them in a secund pass copy the unique one to the actual buffer.
the result is now is much faster