added part tow of the scan prefix sum, and the very good knew is that it take almost zero time,
so small that the profile can' measure it. I was worry that the operation that we will use a lot could become a limiting bother neck, but so far this is very good. scans for an array of 32k elements runs under 4 us, very good news.
we keep going. next make the array of pairs. this is a big one.