cuda is a fucking tour the force is programeing.
the language needs a lot of work from NVidia.
they should leave the work of compile tool to Microsoft or intel, this is quite terrible, form any angle. lot of stuff does not really work as the doc say. the tools leave a lot to be desired.
all in all it is a terrible language for a large scale application.