ok after trying to make a top down scene graph (wip)
it is obviuos that that even top down trees are more tight and probably faster,
it is not practical for GPU, at the end the branch has to be run in one cores wich is really bad for many core.
so I am now experimenting with bottom up construction,
I tried this before and advandone the idea because the result where not any better, but for GPU, they are possible the one few options that can be full parallelized.