fixed crash bug when variance is zero,
which can be if a small bash is at the same position, and this can happen on scene initialization when bodies are added to teh scene, then teh matrix is set at a later frame. The vehicle demo does that for some vehicle internal parts which are places on purpose at the vehicle center.