using proper cuda timer.
still not accuarate but at least is the timer for call to call.

measuring teh time is quite hard because teh cuda drive interrupt the kernels,  so the only way to have a acurate measure of time is to just compute time at teh beginning and end of each kernel call,  which is too much work and an over kill for as debug function.

as it stands now, most time the frame time is the complement of the frame (16.0 - t) ms
but that's better that random value given by clock64