making memory leak tracker account for global allocation
when using globals that allocate memory, they are all tracked.
but in cpp is when a variable is generate in a different module,
it is possible that the deletion of that variable come after the memory leak track was deleted.
this will appear as a memory leak when it is not.