adding the machine learning series of exercises in order of complexity
These are a series of self taught exercises of the foundation of reinforcement learning starting form the very basics
the goal is to learn how to program some of the Q learning method for using it on Physics controllers for Characters.

afte many year of trying zero moment point base characters controllers with very limited success, I decided to give a try to Machine learning techniques base of Temporal Differences.
The Bellman equation seems to be a much better method for dynamics equilibrium than the normal method base of Linear Equation that solve the stability for onle the next step using control theory.
controllers base of inverted pendulum, Dynamaics IK, and other are just too complex, and even when they work, the result is far form desirable.
the best I have seem is robot liek Honda Asimov, an Boston Dynamics Atlas, They are really nice, but requires ton of human resources  and cooperation.
So I will give a try tow Reinforcement learning and see what I can get.
I do no have very high expectation, but it is a new method to attack teh problem.