testing that in the absent  of gravity, motion of limb should preserve linear and angular momentum.
what this means is that when there is no external force and torques,  the movement
of the animation should gives any angular net angular an linearl velocity to the model.

This is violated went using PD to controller to calculate joint torques base of  IK joint angles calculation,

In newton the conservation is valid, for enough simulation frames that the model can be consider accurate.
after many simulations frames doe to the non linearity of the problems, numerical integration error crisp in and the model drift gaining some linear and angular momentum.