doing very first test (conservation of momentum,)
if the body is on teh air, in zero gravity, the leg motion should not add total momentum(linear or angular), to the system
This is a very hard task, with numerical integration since the integration either add or substract momentum to a rigid body.

The gal in that momentum conservation should last for a reasonable number of frames,  and any amount added due to numerical errors, should be compensated by natural damping.

the good knew is that the test since to pass very well.
we can move to teh next step.