trying balance loop (wp)
start to show sign of progress but I have a bug in the IK solver,
having two different solver generate result that are slightly different and that a sort of error.
the results are algebraically equivalent, but the order and the matrix plus internal joint state, causes enough divergent between the IK and generic solver.
therefore I need to consolidate both to eliminate that sort of errors..