using Maximo mocap models
http://mocap.cs.cmu.edu/  need lot of processing
the Bvh formal does not has T pose, and the tool has to guess one.
this is not problem for animation, but when authoring the mesh for physics driven keyframe,
the procedural T poses lack symmetry, thsi could be fix, but need an skillful artist

Maximo file are already converted to fbx format, are easier to manipulate, plus they also has a large set of skinned models.