clean up DQN implementation also trying parallel optimization.
It seem vanilla parallel optimizer is not yielding  any gain, this is because the network are small and the batches also too small, so task switch is a significant cost.
but the good part is that is no slower, so I leave it. Maybe later I optimized it with begin end semantic.