separate softmax activation.
It seems that polgycy gradient mathod use softmax as last activation, but do no use the maximun likelihood gradient, so I seprate the sofmat activations.

sofmax can work with MSE and can be used in any layer
ategorrical softmax that unoy works with maximum likelihood but can onel be used as teh last layer in conjuntion with the Categorical Loss finction.