rolling out first version of pure policy gradient method
vanilla policy gradient implementation
uses a naïve base line but just calculating the montecarlo total reward of a state action,
and applying a Gaussian normalization.

what I do no liek is that the reward tend to be larger that 0ne, which does no make sence for a categorical SoftMax,  but I have not found any documentation that discuses that issue.
I assume that the more advance algorithm deal with that by actually apply a clipping to teh gradient.