best cifar-10 cnn at 79.76% accuracy on test data
It took 14 hours to train this dnn with 140k param.
the cnn is still small as typical gpu configuration go, but I am reaching the limit of what I can do on a cpu.
I will do one last training doubling the param to 250k and see what I get.
them it is upto the GPU support using vulkan.