show desired suport feature (traingle, line of convex poly
next, show location of zero moment point and the distance to the desire suport feature.
the idea is that the RL agant will learn how to move the model body so that it minimizes the distance from the zero moment point to the suport feature. (wip)
so far this method never worked before, I asume that usng linear dunamic, the controller can only predict one step in the future, but it seems that sometime the agent should take one bad step that can maximize some reward in a not so distance future.
using the Belmman equation and deep proxima policy with deep neural net , this seems possible, so I am trying one more time.