coding the policy gradient agent method, (wip)
so far the zero moment point seem to work quite nicely. It is very obvious that for teh most part the zmp and the center of projection of teh COM coincide generation a kinematic stable condition.
but teh interesting part will be if the RL agent can find location when the ZMP and teh COP do not coincide, and the model making the model dynamically stable.
This is important because the interesting locomotion sequence are the one when teh model is dynamically stable. but let us see I fail on this so many time, that I will be happy if the agent find a kinematic stable polycy.