traning quad to move it zmp to teh supoort polygon (wip)
the agent finds the best position but it does at a terrible local minimal.
I think I need to make a better reword function, because once it figure that by moving left and back, it does no need to do anything else so it stops learning.
nevertheless this is a very good result, because the agent learn something.
anyway continue experimenting.