try to make a more phisiocally plausible leg animation.
the martch animation try to keep the feed fix when in air, thsi causes the model to add lot of effort to inforce that condition.
in most causes leg animation retrieve clos to the body com when the leg is lifted, so I will try to add that to teh pose generators. (wip)