adding better baseline value to policy grad method. (wip)
normalizing trajectory does not works too well for very long trajectories where the reward change very little.
the problem is that the vector of reward is almost flat, and the normalization make so that very small reward differences long the path dominate the normalization, when ethe really sould translate to near zero advantages.

using a q value neural net should yield better estimation of the value of the state, so that when the update calculation gradient for those state, the advantage will be almost zero regales of the log prob of teh action. (wip)