more tweaks of policy gradient (wip)
added q value base line estimator using a deep neural net.
so far best quad controller when lifting the first leg.
need to see if if learn when lifting a diffrent leg.
this has been the biggest challenge so far,
the problem has a simple solution, so either the controller is wrong, or the model is not well defined.
I keep trying.