Woo-hoo, finally quad robot learn to lift one leg while maintaining balance.
letting to train for 30 million steps intead of 10 million made a huge difference
the cost is that it take around 5 hours on my old PC.
it seems the secret is a much longer training section.

The good thing is that I can improve the time by using multi agents training on my newer pc. but that's another story.

for now next step is to see if the agent learns to animate more than one leg.