added better sky box algorithm,
nwo onel rendering one quad in clip space.
instead of a full cube in world space.
a full cube is 6 screen size quads, yield lot of rasterization.
so this should be a substantial optimization in GPU rasterization.
not that is a big deal, but it is a more elegant solution.