some adaptive thread optimization (wip)
dividing the job array in equal size chunks allow for a lock and atomic free thread service, which is very nice.
however with new cpu architectures, when they have high performance and low performance core,
the method force make the low performance core the limiting factor since the high performance will comple the tack first and will have to weight for eth lower performance to finish their chunks.

I am now using a atomic CAS counter on most the tread so that the jobs are serve in a first come first served manner, in the pass this has produce thread contention. but I am now using a batch size in order to minimize the number of time the thread hit the atomic counter.
hopefully this is a good compromise between 100 lock free, and some lock but allowing variable tread performcace.

for big little core better thread balaccane