start adding the sgoma optimization of the policy gradients
so far good result by using a const sigman sigma, but for more compex systems with many actions,
the sigma should learned for each action.
example driving a vehicle, we can have a radio or a wind shield actions, this action does not affect driving,
so whe driving the should get a very narrow sigman, while the sigma for sterring, breaks, gear and so on should souel be optimized.