using some type cast before making ndArray 64 bit  (wip)
whan using Deep leaning, some arrays can be larger that few gigabytes, and ndArray of int is not enought to hold the size.
I coule make the size and unsigned but that will onle extendt to 4 gbyte, probably enought but might fail in a verye near future.
I alrady have a 32 gbute PC with a 16 bybyte videos card, so a 32 buit int is not enought.