total waste of time with vulkan.
the only platform I can get working is Nvidia, and if this ii the case,
there is not point on doing vulkan when I can do far better with cuda.
amd driver are buggy as hell, so that is a problem.
Intel vulkan drive are terrible the arch 770 run slower that the single code CPU.