next start adding the cnn  gpu layer
so far gpu seems marginally faster than cpu, in fact I can make cpu equally or probably faster.
But is seem that this is because the model are just too small, for very large model the consensus is that this is where GPU can make a big difference.
for that I will try the Convolutional layer and try 256 x 256 images.