try using diffrent method for controlling the quad robot
I will try the  Joint state vector and oppose to the effector state vector.

Joint state seem a more detail measure of teh model state, than the effector state which is a calculated state.
also the joint state is parametric and can allocate for redundancy, while the effector state does not.

anyway experiment with this and see the results.
the expected result is a convergence in lot of fewer steps,
by using the joints state, a deep neural net should be able to synthesize complex animations animation.