optimized gigmopid and sigmoidLinear activations.
sigmoid limear seeems alsot as good as relu activation.
the fact that does not has a discontinue point, and  reduce vanishig gradient make it a very subtitue for relu.
it is also much simpler than teh Gaussia Relu which similar but expensive to evaluate.
so trying sigmoidLinear and see what we get.