only using Xavier weigh initialization to avoid huge bias
so far it seem that the best method for initializing weight is the xavier method.
that is the bias are set to zero, and the matrix weight are set to a normal distribution
n(0, sqrt(1/n)
when n in the number of neuron in the layer.
this seems to work better even for non tanh activation layers.