trace reward for testing
zmp is too critical, the body can be in dynamics equilibrium and yet be unbalanced. this happens when the zmp project to the diagonal of the support polygon. therefore It need a supplement reward.
one way could be make the support polygon smaller this way the zmp will move inside the polygon instead of the edge.

another maybe a penalty for not having the desired support in contact with the floor. but that could be too unreliable.