adding holonomic test to close joint solver ik chains. (wip)
when an IK chain loses one or more degree of freedom, the system become over determine and teh constraint solver use the penalties to make holonomic, generating extreme internal force and torques.
Thsi can be tolerated for a few frame, but after a while this become a positive feedback that degenerates in a large floating point error that lead to an explosion.

ther are few solution for this,
one will be to modify the constraint solver so that the holonomic condition is satisfy, but this is quite difficult, in fact one of the solution is adding constraint regularization, which I am already doing.
constraint regularization can only to sustain for a few frames before the system degenerates.

In general this condition raretly happen, since system are designed for this condition to no happens for a long period of time.

 however when doing deep learning training, this condition does happens, simply because teh stochasticity of the training so I need to come up with a solutions.

my solution is to simply detect teh lose of a degree of freedom and training the system to avoid steps that avoid that condition.  hopefully that is enough.