testing arm reaching robot
big different between the ai trained and the Inver dynamics trained.
the Inver dynamics can't check it the arm in going to violate holonomicity  between effector and target
went the erro is too large.

the ai trained seem no have have that problem since it always get to the target by generation a trajectory in which holonomicity is valid at each time steps.

this alone justify the effro investe in deep lerning.