adding all 6 degrees of freedom, to simple industrial robot.
the robot reach teh target very nicely, but is has the problem that is does not knwo how comple is the work space. so it self collides and and interest
I am hopping that a deep neural ne can figure out that comple work space of the robot.
so let us try thsi again.