make min sigma small
It is possible that a very critical robot has to reduce exploration a lot in some cases.
the double pendulum demos show that when a the center and small random change can cause the to to topple.
the algorism should reduce sigma to the minimum in those cases. at lest that's the idea.
let use see how it does.