start adding a q base optimized.
While policy gradient methods tend to be more stable than Q value based methods,
they require collecting a large amount of data, which can be impractical on hardware with limited resources.
As model complexity grows from medium to large scale, the variance in the collected data increases exponentially.
This happens because uncertainty in the data stems from all control outputs,
and any moderately complex robot typically deals with an input vector composed of dozens of signals.

To manage this high variance, a common solution is to train thousands of agents in parallel.
This approach is motivated by the theoretical foundation of policy gradient methods,
which update model parameters based on the expected reward over all possible trajectories.
However, in practice, we must estimate this expectation from a finite set of sampled trajectories.
The challenge arises because, while our ability to generate data grows linearly
(or sub linearly) with system resources,
the variance of that data increases exponentially with the size of the action vector.
This becomes a major bottleneck for small to medium sized systems.

In fact, even a high end single GPU system may struggle with this demand,
which often limits the practical use of policy gradient methods to organizations with access
to large scale computing resources such as supercomputers plus al large number of humans in the loop to supervise data generation.

therefore, it is my opinion and conclusion that Q base methos are more sub table for medium small systems.

I am nwo start to write a sof actor critic trainned (wip)