adding the double multiple critic version (wip)
the multiple critic version of DDPG is 66% slower than the DDPG
but reduce the variance a lot.
I have not seem the benefic of the delay so I am omitting that part.

one benefic of the DDPG is that is is way of a lot more sample efficient than the on policy methods.
I have to investigate Soft Actor critic, so that I can add more robustness for harder problems..