tune vehicle now
that we have a stronger solver, the vehicle tire contact model act as the book say.
for the first time I see that in fact wheel vehicle, over stee more than for and front wheel drive.
This is the first to I see this since the original newton that used a conjugate gradient solver, but was just too slow. Now that I have a better solver that is more accurate but a way faster, I need to go over a recalibration.